// frontend/src/app/models/ml-pipeline.models.ts
export enum VisualizationType {
  FEATURE_IMPORTANCE = 'feature_importance',
  CONFUSION_MATRIX = 'confusion_matrix',
  ROC_CURVE = 'roc_curve',
  REGRESSION_PLOT = 'regression_plot',
  DECISION_TREE = 'decision_tree',
  PREDICTIONS_DISTRIBUTION = 'predictions_distribution',
  PERFORMANCE_TIMELINE = 'performance_timeline',
  METRICS = 'metrics'
}